Address Transactions Fix
